Two changes to IC: random subset of modes and evenly fitting modes.
Only a fraction of the modes within the wavenumber bounds [kmin, kmax]
are excited.  This reduces init time while still allowing
for broad spectrum ICs.

Only modes that fit exactly within the box are excited.  This
is controlled by letting the x,y extents be a power of 2 and
then setting kscal to (the opposite) of that power.  For example,
if extent_mesh[0] = extent_mesh[1] = 1/4, then kscal = 4.